+++ illyriacore/src/main/java/eu/illyrion/illyriacore/interfaces/ItemsInterface.java
@@ -1,4 +1,4 @@
-package eu.illyrion.illyriacore.configs;
+package eu.illyrion.illyriacore.interfaces;
 
 public interface ItemsInterface {
     // folder
